🚚 Move homepage in another file
(redirect logic in index page)
