+++ front/app/index.tsx
@@ -1,13 +1,7 @@
-import { ScrollView } from "@gluestack-ui/themed";
+import { router } from "expo-router";
 
-import Header from "../components/Header";
-import OfferList from "../components/offers/OfferList";
+export default function Index() {
+  router.replace("/home");
 
-export default function Home() {
-  return (
-    <ScrollView w="$full" minHeight="$full" bgColor="$white">
-      <Header title="👋 Welcome" search={{ placeholder: "Search an offer" }} />
-      <OfferList title="Recommended for you" />
-    </ScrollView>
-  );
+  return null;
 }
